refactor(repository-hazelcast): rename PluginConfiguration to RateLimitRepositoryConfiguration
